+++ src/components/pages/dispenser/index.tsx
@@ -152,11 +152,10 @@ const Dispenser: FC<ReduxType> = ({
       loading={loading}
       onClose={() => toggleUpdateLinksPopup(false)}
       onSubmit={links => {
-        
         if (!dispenser_id) { return alert('Dispenser ID not found') }
         if (!encrypted_multiscan_qr_enc_code) { return alert('encrypted_multiscan_qr_enc_code not found') }
+        if (links_count === undefined) { return alert('links_count not found') }
         if (!links) { return alert('Links not found') }
-
         addLinksToQR(dispenser_id, links, encrypted_multiscan_qr_enc_code, links_count, () => {
           toggleUpdateLinksPopup(false)
         })