+++ src/configs/address-specific-options.tsx
@@ -27,7 +27,9 @@ export const addressSpecificOptions: TAddressConfig = {
     betaFeaturesAvailable: ['dispenser']
   },
   '0xb4c3d57327d4fc9bcc3499963e21db1a5435d537': COINBASE_CONFIG,
-  '0x3e12f31db4f1250e077171568263d0856d897428': COINBASE_CONFIG,
+  '0x3e12f31db4f1250e077171568263d0856d897428': {
+    betaFeaturesAvailable: ['dispenser']
+  },
   '0x0553ada5829184a7818dc866367d77334183603e': {
     betaFeaturesAvailable: ['dispenser']
   },
@@ -38,6 +40,9 @@ export const addressSpecificOptions: TAddressConfig = {
     betaFeaturesAvailable: ['dispenser']
   },
   '0xe2f91e21c6689382c1c3dd8cd70633c39c911c4a': {
+    betaFeaturesAvailable: ['dispenser']
+  },
+  '0xb7d98f28d9f7ef2732802c62bcac12c1b0d5edfb': {
     betaFeaturesAvailable: ['dispenser']
   }
 }