+++ src/helpers/format-time.tsx
@@ -1,10 +1,10 @@
-import { TFormatDate } from 'types'
+import { TFormatTime } from 'types'
 
 const format = (value: number) => {
   return value < 10 ? `0${value}` : value
 }
 
-const formatTime: TFormatDate = (date) => {
+const formatTime: TFormatTime = (date, showSeconds = false) => {
   const dateObj = new Date(date)
   const hours = dateObj.getHours()
   const minutes = dateObj.getMinutes()
@@ -12,7 +12,10 @@ const formatTime: TFormatDate = (date) => {
   const minutesFormatted = format(minutes)
   const secondsFormatted = format(seconds)
   const hoursFormatted = format(hours)
-  return ` ${hoursFormatted}:${minutesFormatted}:${secondsFormatted}`
+  if (!showSeconds) {
+    return `${hoursFormatted}:${minutesFormatted}`
+  }
+  return `${hoursFormatted}:${minutesFormatted}:${secondsFormatted}`
 }
 
 export default formatTime